requirements.txt
aiogram
requests
moviepy
opencv-python
numpy
yt-dlp
google-auth
google-auth-oauthlib
google-api-python-client
python-dotenv
